# backend/tests/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnboardingCompleteRequest": {
        "properties": {
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          }
        },
        "required": [
          "assessment_id"
        ],
        "title": "OnboardingCompleteRequest",
        "type": "object"
      },
      "OnboardingCompleteResponse": {
        "properties": {
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          }
        },
        "required": [
          "onboarding_completed"
        ],
        "title": "OnboardingCompleteResponse",
        "type": "object"
      },
      "OnboardingRepoRequest": {
        "properties": {
          "github_token": {
            "title": "Github Token",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          }
        },
        "required": [
          "repo_url",
          "github_token"
        ],
        "title": "OnboardingRepoRequest",
        "type": "object"
      },
      "OnboardingRepoResponse": {
        "properties": {
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "verified": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerifiedRepo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "assessment_id",
          "repo_url"
        ],
        "title": "OnboardingRepoResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifiedRepo": {
        "description": "Display-only metadata the SPA shows on the connect-success card.",
        "properties": {
          "default_branch": {
            "title": "Default Branch",
            "type": "string"
          },
          "permissions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "repo_name": {
            "title": "Repo Name",
            "type": "string"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "repo_name",
          "visibility",
          "default_branch"
        ],
        "title": "VerifiedRepo",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/onboarding/complete": {
      "post": {
        "description": "Mark onboarding as complete once the first assessment finishes.",
        "operationId": "complete_onboarding_api_onboarding_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingCompleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/onboarding/repo": {
      "post": {
        "description": "Register a repo and kick off the initial assessment.",
        "operationId": "connect_repo_api_onboarding_repo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRepoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingRepoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Repo",
        "tags": [
          "onboarding"
        ]
      }
    },
...
  }
}